Revert "Sw 96 workout page format"
