+++ SmartWeights/SmartWeights/post-workout/PostWorkout.swift
@@ -7,25 +7,17 @@
 
 import SwiftUI
 
-///structure to display the overall workout progress
-
 struct PostWorkout: View {
-    //the view to show all their overall workout
-    //will also allow them to select data from a certain date
+    
+
+    
     var body: some View {
         ZStack{
             VStack{
                 //back arrow for future implementation
                 HStack{
-                    Button(action: { // Back Arrow
-                        print("Button tapped (temp)")
-                    }) {
-                        Image(systemName: "arrow.left")
-                            .imageScale(.large)
-                            .foregroundColor(.black)
-                            .padding(.leading)
-                        Spacer()
-                    }
+                    Image(systemName: "arrow.backward")
+                        .padding(.trailing, 350)
                 }
                     
                 //Title of the page
@@ -36,12 +28,15 @@ struct PostWorkout: View {
                         .bold()
                         .fontDesign(.monospaced)
                         .multilineTextAlignment(.center)
+
+                 
                 }
                 
                 //calender component
                 //need to export date to get data the correct day
                 
                 Calendar()
+                
                 //display data for that day
                 ZStack(alignment: .leading){
                     Rectangle()
